# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba Resuelve",
        "description": "Prueba Resuelve",
        "contact": {
            "email": "luisvl13@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/equipos": {
            "get": {
                "tags": [
                    "Equipos"
                ],
                "summary": "Obtener lista de Equipos",
                "operationId": "d29638a9b7c4317a2acd4a09086dfa86",
                "responses": {
                    "200": {
                        "description": "Lista de los equipos y sus niveles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "messages": {
                                            "type": "string",
                                            "example": "Operación realizada con exito"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Resuelve FC"
                                                    },
                                                    "niveles": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number",
                                                                    "example": "1"
                                                                },
                                                                "nivel": {
                                                                    "type": "string",
                                                                    "example": "A"
                                                                },
                                                                "goles_mes": {
                                                                    "type": "number",
                                                                    "example": "5"
                                                                },
                                                                "equipos_id": {
                                                                    "type": "integer",
                                                                    "example": "1"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "404"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Equipos"
                ],
                "summary": "Crea un equipo",
                "operationId": "d7039c7ff01f1cfdb5320c1b5fd72ea7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "Resuelve FC"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Equipo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "messages": {
                                            "type": "string",
                                            "example": "Registro creado"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Resuelve FC"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "400"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Conflicto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "500"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}